+++ app/src/main/java/com/tick/taku/notificationwatcher/domain/repository/NotificationRepository.kt
@@ -1,0 +1,12 @@
+package com.tick.taku.notificationwatcher.domain.repository
+
+import android.service.notification.StatusBarNotification
+
+interface NotificationRepository {
+
+    /**
+     * Save notification to db.
+     */
+    suspend fun saveNotification(sbn: StatusBarNotification)
+
+}
